feat: allow user fusion with payment users
